refactor(sdjwtvc): regroup files to prepare for sdjwt/sdjwtvc split
Step 1 of docs/sdjwt-package-split-plan.md. Pure code motion within
package sdjwtvc, no renames or behavior changes: split sdjwtvc.go,
verifier_helpers.go, path_pointer.go, issuer.go and verify.go into the
file layout the eventual eudi/sdjwt + eudi/credentials/sdjwtvc split
will use (disclosure.go, serialization.go, claims.go, builder.go,
processed.go, verify_disclosures.go, payload.go), so the actual
package split in a later step is a mechanical file move.
