+++ eudi/credentials/sdjwtvc/claims.go
@@ -1,0 +1,18 @@
+package sdjwtvc
+
+const (
+	Key_Subject         string = "sub"
+	Key_ExpiryTime      string = "exp"
+	Key_IssuedAt        string = "iat"
+	Key_Issuer          string = "iss"
+	Key_Sd              string = "_sd"
+	Key_SdAlg           string = "_sd_alg"
+	Key_Confirmationkey string = "cnf"
+	Key_NotBefore       string = "nbf"
+	Key_Typ             string = "typ"
+	Key_X5c             string = "x5c"
+	Key_Kid             string = "kid"
+	Key_Ellipsis        string = "..."
+
+	KbJwtTyp string = "kb+jwt"
+)
