+++ eudi/credentials/sdjwtvc/processed.go
@@ -1,9 +1,69 @@
 package sdjwtvc
 
 import (
+	"encoding/json"
 	"fmt"
 	"reflect"
+	"slices"
 )
+
+type ProcessedSdJwtPayload map[string]any
+
+// MarshalJSON ensures that the JSON encoding of the ProcessedSdJwtPayload is deterministic by sorting map keys, which is necessary for consistent hashing of the payload.
+// In order to calculate the hash consistently, the entire payload structure has to be sorted.
+// Fortunately, ProcessedSdJwtPayload is built up from map[string]any structures (where any is either a scalar value, a map, or an array), which we can sort by marshalling to JSON, which already sorts map keys.
+func (p *ProcessedSdJwtPayload) MarshalJSON() ([]byte, error) {
+	p.Sort()
+	return json.Marshal(map[string]any(*p))
+}
+
+// Sort sorts the ProcessedSdJwtPayload in place, by sorting all arrays by their values (when the array is of a scalar type).
+// This ensures that the JSON encoding of the payload is deterministic, which is necessary for consistent hashing of the payload.
+// As the map is keyed, it cannot be sorted itself, but this is handled by the JSON marshalling.
+func (p *ProcessedSdJwtPayload) Sort() {
+	for _, v := range *p {
+		rt := reflect.TypeOf(v)
+		switch rt.Kind() {
+		case reflect.Map:
+			m, ok := v.(ProcessedSdJwtPayload)
+			if ok {
+				m.Sort()
+			} else {
+				panic(fmt.Errorf("unexpected map type in ProcessedSdJwtPayload: %v", rt))
+			}
+		case reflect.Slice, reflect.Array:
+			kind := rt.Elem().Kind()
+			switch kind {
+			case reflect.Float32:
+				slices.Sort(v.([]float32))
+			case reflect.Float64:
+				slices.Sort(v.([]float64))
+			case reflect.Uint8:
+				slices.Sort(v.([]uint8))
+			case reflect.Uint16:
+				slices.Sort(v.([]uint16))
+			case reflect.Uint32:
+				slices.Sort(v.([]uint32))
+			case reflect.Uint64:
+				slices.Sort(v.([]uint64))
+			case reflect.Uint:
+				slices.Sort(v.([]uint))
+			case reflect.Int8:
+				slices.Sort(v.([]int8))
+			case reflect.Int16:
+				slices.Sort(v.([]int16))
+			case reflect.Int32:
+				slices.Sort(v.([]int32))
+			case reflect.Int64:
+				slices.Sort(v.([]int64))
+			case reflect.Int:
+				slices.Sort(v.([]int))
+			case reflect.String:
+				slices.Sort(v.([]string))
+			}
+		}
+	}
+}
 
 // GetClaimValue resolves a claims path pointer against the processed SD-JWT payload,
 // as defined in Appendix C of the OpenID4VCI specification.
