+++ eudi/credentials/sdjwtvc/verify.go
@@ -5,15 +5,11 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"maps"
-	"reflect"
 	"slices"
 	"time"
 
-	jwtOld "github.com/golang-jwt/jwt/v4"
 	"github.com/lestrrat-go/jwx/v3/jwa"
 	"github.com/lestrrat-go/jwx/v3/jwk"
-	"github.com/lestrrat-go/jwx/v3/jws"
 	"github.com/lestrrat-go/jwx/v3/jwt"
 	"github.com/privacybydesign/irmago/eudi/credentials/statuslist"
 	eudi_jwt "github.com/privacybydesign/irmago/eudi/jwt"
@@ -87,64 +83,6 @@ func CreateDefaultVerificationContext(trustedChain []byte) SdJwtVcVerificationCo
 		},
 		Clock:       eudi_jwt.NewSystemClock(),
 		JwtVerifier: NewJwxJwtVerifier(),
-	}
-}
-
-type ProcessedSdJwtPayload map[string]any
-
-// MarshalJSON ensures that the JSON encoding of the ProcessedSdJwtPayload is deterministic by sorting map keys, which is necessary for consistent hashing of the payload.
-// In order to calculate the hash consistently, the entire payload structure has to be sorted.
-// Fortunately, ProcessedSdJwtPayload is built up from map[string]any structures (where any is either a scalar value, a map, or an array), which we can sort by marshalling to JSON, which already sorts map keys.
-func (p *ProcessedSdJwtPayload) MarshalJSON() ([]byte, error) {
-	p.Sort()
-	return json.Marshal(map[string]any(*p))
-}
-
-// Sort sorts the ProcessedSdJwtPayload in place, by sorting all arrays by their values (when the array is of a scalar type).
-// This ensures that the JSON encoding of the payload is deterministic, which is necessary for consistent hashing of the payload.
-// As the map is keyed, it cannot be sorted itself, but this is handled by the JSON marshalling.
-func (p *ProcessedSdJwtPayload) Sort() {
-	for _, v := range *p {
-		rt := reflect.TypeOf(v)
-		switch rt.Kind() {
-		case reflect.Map:
-			m, ok := v.(ProcessedSdJwtPayload)
-			if ok {
-				m.Sort()
-			} else {
-				panic(fmt.Errorf("unexpected map type in ProcessedSdJwtPayload: %v", rt))
-			}
-		case reflect.Slice, reflect.Array:
-			kind := rt.Elem().Kind()
-			switch kind {
-			case reflect.Float32:
-				slices.Sort(v.([]float32))
-			case reflect.Float64:
-				slices.Sort(v.([]float64))
-			case reflect.Uint8:
-				slices.Sort(v.([]uint8))
-			case reflect.Uint16:
-				slices.Sort(v.([]uint16))
-			case reflect.Uint32:
-				slices.Sort(v.([]uint32))
-			case reflect.Uint64:
-				slices.Sort(v.([]uint64))
-			case reflect.Uint:
-				slices.Sort(v.([]uint))
-			case reflect.Int8:
-				slices.Sort(v.([]int8))
-			case reflect.Int16:
-				slices.Sort(v.([]int16))
-			case reflect.Int32:
-				slices.Sort(v.([]int32))
-			case reflect.Int64:
-				slices.Sort(v.([]int64))
-			case reflect.Int:
-				slices.Sort(v.([]int))
-			case reflect.String:
-				slices.Sort(v.([]string))
-			}
-		}
 	}
 }
 
@@ -406,56 +344,6 @@ func (v *sdJwtVcProcessor) verifyTimeFields(issuerSignedJwtPayload *IssuerSigned
 	return nil
 }
 
-func parseSdField(value any) ([]HashedDisclosure, error) {
-	strs, ok := value.([]any)
-	if !ok {
-		return []HashedDisclosure{}, fmt.Errorf("failed to convert _sd field to []any (%s)", value)
-	}
-	if len(strs) == 0 {
-		return []HashedDisclosure{}, fmt.Errorf("when the _sd field is present it may not be empty")
-	}
-	result := []HashedDisclosure{}
-	for _, s := range strs {
-		sStr, ok := s.(string)
-		if !ok {
-			return []HashedDisclosure{}, fmt.Errorf("failed to convert value in _sd array to string (%v)", s)
-		}
-		result = append(result, HashedDisclosure(sStr))
-	}
-	return result, nil
-}
-
-func parseConfirmField(value any) (*CnfField, error) {
-	anyMap, ok := value.(map[string]any)
-	if !ok {
-		return nil, fmt.Errorf("failed to parse as anymap: %v", value)
-	}
-
-	// We support jwk and kid (with did:jwk method) confirmations.
-	jwkAny, ok := anyMap["jwk"]
-	if ok {
-		jwkJson, err := json.Marshal(jwkAny)
-		if err != nil {
-			return nil, err
-		}
-		key, err := jwk.ParseKey(jwkJson)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse key (%v) from json: %v", value, err)
-		}
-		return &CnfField{Jwk: &key}, nil
-	}
-	kidAny, ok := anyMap["kid"]
-	if ok {
-		kidStr, ok := kidAny.(string)
-		if !ok {
-			return nil, fmt.Errorf("failed to parse kid field as string: %v", kidAny)
-		}
-		return &CnfField{Kid: &kidStr}, nil
-	}
-
-	return nil, fmt.Errorf("failed to parse cnf field: unsupported confirmation method, expected jwk or did:jwk: %v", value)
-}
-
 // parseStatusClaim reads the `status` claim from a verified jwt.Token
 // into the structured form expected by the Token Status List
 // pipeline. Only the `status_list` member is parsed in v1; other
@@ -511,205 +399,6 @@ func parseStatusClaim(token jwt.Token) (*statuslist.StatusClaim, error) {
 		return nil, fmt.Errorf("status_list.idx is not a number: %T", idxRaw)
 	}
 	return &statuslist.StatusClaim{StatusList: &statuslist.Reference{Index: idx, URI: uri}}, nil
-}
-
-func verifyAndProcessDisclosures(sdAlg iana.HashingAlgorithm,
-	issuerSignedJwtClaims *map[string]any,
-	disclosures []EncodedDisclosure,
-) (ProcessedSdJwtPayload, []*DisclosureContent, error) {
-	// Step 3.a: decode all disclosures and calculate their digests
-	decodedDisclosuresMap := make(map[HashedDisclosure]*DisclosureContent, len(disclosures))
-	for _, disc := range disclosures {
-		decodedDisclosure, err := DecodeDisclosure(disc)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to decode disclosure: %v", err)
-		}
-
-		digest, err := HashEncodedDisclosure(sdAlg, disc)
-		if err != nil {
-			return nil, nil, fmt.Errorf("failed to hash disclosure: %v", err)
-		}
-
-		decodedDisclosuresMap[digest] = &decodedDisclosure
-	}
-
-	// Keep a list of all disclosures for return value
-	decodedDisclosures := slices.Collect(maps.Values(decodedDisclosuresMap))
-
-	// Step 3.b - 3.e: Identify all digests in the Issuer-Signed JWT recursively and replace them with the actual disclosure values
-	err := processEmbeddedDisclosures(issuerSignedJwtClaims, decodedDisclosuresMap)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Step 4: double encountered digests are signalled as soon as they are found during processing
-
-	// Step 5: if a disclosure was not referenced (i.e. removed from the map), the SD-JWT is invalid
-	for _, disclosure := range decodedDisclosuresMap {
-		if !disclosure.IsTouched() {
-			return nil, nil, fmt.Errorf("one or more disclosures were not referenced in the issuer signed jwt")
-		}
-	}
-
-	// Step 3.f: remove the _sd_alg field from the claims
-	if issuerSignedJwtClaims == nil {
-		issuerSignedJwtClaims = &map[string]any{}
-	}
-	delete(*issuerSignedJwtClaims, Key_SdAlg)
-
-	return ProcessedSdJwtPayload(*issuerSignedJwtClaims), decodedDisclosures, nil
-}
-
-func processEmbeddedDisclosures(claims *map[string]any, decodedDisclosures map[HashedDisclosure]*DisclosureContent) error {
-	// Only process if there are any claims
-	if claims == nil {
-		return nil
-	}
-
-	// Start with _sd field first
-	err := processSdClaim(claims, decodedDisclosures)
-	if err != nil {
-		return err
-	}
-
-	// Then process all other claims
-	for claimKey, claimValue := range *claims {
-		// If the property is a nested object, recursively process it
-		if claimMap, ok := claimValue.(map[string]any); ok {
-			err := processEmbeddedDisclosures(&claimMap, decodedDisclosures)
-			if err != nil {
-				return err
-			}
-			(*claims)[claimKey] = claimMap
-			continue
-		}
-
-		// Or, of it is an array, process each element
-		if arrayValue, ok := claimValue.([]any); ok {
-			processedArray := []any{}
-			for _, arrayElemValue := range arrayValue {
-				// Check if the value is a disclosure (format should be {"...":"<digest>"}) or not, and if so,
-				// verify the array element disclosure exists
-				if valMap, ok := arrayElemValue.(map[string]any); ok {
-					if arrayElemDisclosureDigestVal, ok := valMap[Key_Ellipsis]; ok {
-						// It's an embedded disclosure digest...
-						arrayElemDisclosureDigestStr, ok := arrayElemDisclosureDigestVal.(string)
-						if !ok {
-							return fmt.Errorf(
-								"array element, which should be an embedded disclosure digest, is not a valid digest: %v",
-								arrayElemDisclosureDigestStr,
-							)
-						}
-
-						disclosureDigest := HashedDisclosure(arrayElemDisclosureDigestStr)
-						if embeddedDisclosure, ok := decodedDisclosures[disclosureDigest]; ok {
-							// Check for array element validity (i.e. should be in format ["...": "<digest>"])
-							if embeddedDisclosure.IsTouched() {
-								return fmt.Errorf(
-									"digest %s has been referenced multiple time in the SD-JWT",
-									disclosureDigest,
-								)
-							}
-							if !embeddedDisclosure.IsArrayElement() {
-								return fmt.Errorf(
-									"embedded disclosure %s is expected to be an array element, but is not",
-									embeddedDisclosure.Key,
-								)
-							}
-
-							// Otherwise, replace the array element with the actual value from the disclosure
-							embeddedDisclosure.Touch()
-							processedArray = append(processedArray, embeddedDisclosure.Value)
-						}
-
-						// In case no disclosure is found for the digest; the value will be ignored (potential decoy digest)
-						// Either way; we can continue to the next array element
-						continue
-					}
-
-					// Complex value, but no embedded digest: just copy it
-					// Recursively process the claim map to find further embedded digests
-					err := processEmbeddedDisclosures(&valMap, decodedDisclosures)
-					if err != nil {
-						return err
-					}
-					processedArray = append(processedArray, valMap)
-				} else {
-					// Simple value, just copy it
-					processedArray = append(processedArray, arrayElemValue)
-				}
-			}
-
-			(*claims)[claimKey] = processedArray
-			continue
-		}
-
-		// No embedded disclosures found, just copy the claim as is
-		(*claims)[claimKey] = claimValue
-	}
-
-	return nil
-}
-
-func processSdClaim(claims *map[string]any, decodedDisclosures map[HashedDisclosure]*DisclosureContent) error {
-	// Only process if there are any claims
-	if claims == nil {
-		return nil
-	}
-
-	// Check if there's an _sd field at this level
-	sdValue, ok := (*claims)[Key_Sd]
-	if !ok {
-		return nil
-	}
-
-	// Found disclosure digests at this level.. replace with disclosure values
-	sdDigests, err := parseSdField(sdValue)
-	if err != nil {
-		return fmt.Errorf("failed to parse digests for claim %q: %v", Key_Sd, err)
-	}
-
-	for _, sdDigest := range sdDigests {
-		// Disclosure cannot be found for digest; ignore the digest
-		if embeddedDisclosure, ok := decodedDisclosures[sdDigest]; ok {
-			if embeddedDisclosure.IsTouched() {
-				return fmt.Errorf("digest %s has been referenced multiple time in the SD-JWT", sdDigest)
-			}
-			if embeddedDisclosure.IsArrayElement() {
-				return fmt.Errorf("embedded disclosure %s appears to be an array element, which is not expected here", embeddedDisclosure.Key)
-			}
-			if embeddedDisclosure.Key == Key_Sd {
-				return fmt.Errorf("embedded disclosure %s has an `_sd` field, which is not allowed", embeddedDisclosure.Key)
-			}
-			if embeddedDisclosure.Key == Key_Ellipsis {
-				return fmt.Errorf("embedded disclosure %s has an `...` field, which is not allowed", embeddedDisclosure.Key)
-			}
-			if _, ok := (*claims)[embeddedDisclosure.Key]; ok {
-				return fmt.Errorf("embedded disclosure key %q already exists at this level", embeddedDisclosure.Key)
-			}
-
-			embeddedDisclosure.Touch()
-			(*claims)[embeddedDisclosure.Key] = embeddedDisclosure.Value
-		}
-	}
-
-	// Delete the _sd field after processing
-	delete(*claims, Key_Sd)
-
-	return nil
-}
-
-func extractClaimsAndDisclosuresDigestsFromToken(token jwt.Token) (map[string]any, error) {
-	issuerSignedJwtClaims := map[string]any{}
-	for _, key := range token.Keys() {
-		var value any
-		if err := token.Get(key, &value); err != nil {
-			return nil, fmt.Errorf("failed to get extra claim %s: %v", key, err)
-		}
-
-		issuerSignedJwtClaims[key] = value
-	}
-	return issuerSignedJwtClaims, nil
 }
 
 // Decode the JWT into a token, verify the signature with the X.509 certificate in the header and verify the certificate is trusted (both against root/intermediate certs and CRLs).
@@ -917,16 +606,6 @@ func (v *verifierKeyBindingProcessor) parseAndVerifyKeyBindingJwt(
 	return &payload, nil
 }
 
-func decodeJwtWithoutCheckingSignature(jwtString string) (header map[string]any, claims map[string]any, err error) {
-	parser := jwtOld.NewParser()
-	var claimsResult jwtOld.MapClaims
-	token, _, err := parser.ParseUnverified(jwtString, &claimsResult)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to parse JWT: %v", err)
-	}
-	return token.Header, claimsResult, err
-}
-
 // ============================= Holder processing =====================================
 
 type HolderVerificationProcessor struct {
@@ -957,20 +636,4 @@ func (p *holderVerifierKeyBindingProcessor) ProcessAndVerifyKeyBindingJwt(kbjwt 
 // ParseAndVerifySdJwtVc is used to verify an SD-JWT VC using the verification options passed via the context parameter.
 func (v *HolderVerificationProcessor) ParseAndVerifySdJwtVc(sdjwtvc SdJwtVcKb) (*VerifiedSdJwtVc, error) {
 	return v.sdJwtVcProcessor.ProcessAndVerifySdJwtVc(sdjwtvc, &holderVerifierKeyBindingProcessor{})
-}
-
-// ========================================================================
-
-type JwtVerifier interface {
-	Verify(jwt string, key any, sigAlg jwa.SignatureAlgorithm) (payload []byte, err error)
-}
-
-type JwxJwtVerifier struct{}
-
-func NewJwxJwtVerifier() *JwxJwtVerifier {
-	return &JwxJwtVerifier{}
-}
-
-func (v *JwxJwtVerifier) Verify(jwtString string, keyAny any, sigAlg jwa.SignatureAlgorithm) (payload []byte, err error) {
-	return jws.Verify([]byte(jwtString), jws.WithKey(sigAlg, keyAny))
 }
